Code to support layout animation, image fading, image color tinting and opacity
